+++ lib/renderer/api/ipc-renderer.js
@@ -5,6 +5,41 @@ const v8Util = process.atomBinding('v8_util')
 
 // Created by init.js.
 const ipcRenderer = v8Util.getHiddenValue(global, 'ipc')
-require('./ipc-renderer-setup')(ipcRenderer, binding)
+
+ipcRenderer.send = function (...args) {
+  return binding.send('ipc-message', args)
+}
+
+ipcRenderer.sendSync = function (...args) {
+  return JSON.parse(binding.sendSync('ipc-message-sync', args))
+}
+
+ipcRenderer.sendToHost = function (...args) {
+  return binding.send('ipc-message-host', args)
+}
+
+ipcRenderer.sendTo = function (webContentsId, channel, ...args) {
+  if (typeof webContentsId !== 'number') {
+    throw new TypeError('First argument has to be webContentsId')
+  }
+
+  ipcRenderer.send('ELECTRON_BROWSER_SEND_TO', false, webContentsId, channel, ...args)
+}
+
+ipcRenderer.sendToAll = function (webContentsId, channel, ...args) {
+  if (typeof webContentsId !== 'number') {
+    throw new TypeError('First argument has to be webContentsId')
+  }
+
+  ipcRenderer.send('ELECTRON_BROWSER_SEND_TO', true, webContentsId, channel, ...args)
+}
+
+const removeAllListeners = ipcRenderer.removeAllListeners.bind(ipcRenderer)
+ipcRenderer.removeAllListeners = function (...args) {
+  if (args.length === 0) {
+    throw new Error('Removing all listeners from ipcRenderer will make Electron internals stop working.  Please specify a event name')
+  }
+  removeAllListeners(...args)
+}
 
 module.exports = ipcRenderer
